chore: extract formatValue and isValidAmount to shared
These two number-formatting helpers in app/scripts/lib/util.ts are
pure utilities consumed only by UI percentage-overview components.
Moving them to shared/lib/format-value lets UI import them without
crossing the architectural boundary; app/scripts/lib/util keeps
re-exporting them for any background callers.
