+++ app/scripts/lib/util.ts
@@ -56,6 +56,7 @@ export {
   isValidEmail,
   isWebOrigin,
 } from '../../../shared/lib/url-utils';
+export { formatValue, isValidAmount } from '../../../shared/lib/format-value';
 
 /**
  * Minimal type for User-Agent Client Hints API (NavigatorUAData).
@@ -535,24 +536,6 @@ export function formatTxMetaForRpcResult(
   }
 
   return formattedTxMeta;
-}
-
-export const isValidAmount = (amount: number | null | undefined): boolean =>
-  amount !== null && amount !== undefined && !Number.isNaN(amount);
-
-export function formatValue(
-  value: number | null | undefined,
-  includeParentheses: boolean,
-): string {
-  if (!isValidAmount(value)) {
-    return '';
-  }
-
-  const numericValue = value as number;
-  const sign = numericValue >= 0 ? '+' : '';
-  const formattedNumber = `${sign}${numericValue.toFixed(2)}%`;
-
-  return includeParentheses ? `(${formattedNumber})` : formattedNumber;
 }
 
 type MethodData = {
