+++ ui/components/app/wallet-overview/aggregated-percentage-overview.tsx
@@ -13,9 +13,7 @@ import { getPreferences } from '../../../../shared/lib/selectors/preferences';
 import { getSelectedInternalAccount } from '../../../../shared/lib/selectors/accounts';
 import { getCurrentChainId } from '../../../../shared/lib/selectors/networks';
 import { useAccountTotalFiatBalance } from '../../../hooks/useAccountTotalFiatBalance';
-// TODO: Remove restricted import
-// eslint-disable-next-line import-x/no-restricted-paths
-import { formatValue, isValidAmount } from '../../../../app/scripts/lib/util';
+import { formatValue, isValidAmount } from '../../../../shared/lib/format-value';
 import { getIntlLocale } from '../../../ducks/locale/locale';
 import {
   Display,
